add support for non-composite sites (toss a warning)

+++ util/validateCompositeSites.sh
@@ -37,8 +37,8 @@ debug=0
 # read commandline args
 while [[ "$#" -gt 0 ]]; do
 	case $1 in
-	    '-q') quiet="1"; shift 0;;
-	    '-X') debug="1"; shift 0;;
+		'-q') quiet="1"; shift 0;;
+		'-X') debug="1"; shift 0;;
 		*) checkurls="${checkurls} $1"; shift 0;; # URLs to read
 	esac
 	shift 1
@@ -70,106 +70,123 @@ log ()
 logerr ()
 {
   if [[ $quiet == 0 ]]; then 
-    echo -e "$2"
+	echo -e "$2"
   else
-    echo -e "$1$2"
+	echo -e "$1$2"
   fi
 }
 
 norm="\033[0;39m"
 green="\033[1;32m"
 yellow='\e[0;30m\e[1;43m' # yellow reversed
+blue="\033[1;34m"
 red="\033[1;31m"
 OK=0
 notOK=0
 
 checkCompositeXML ()
 {
-    # parse the file for <child location="" URLs>
-    checkurl=$1
-    indentNum=$2
+	# parse the file for <child location="" URLs>
+	checkurl=$1
+	indentNum=$2
 	if [[ $debug -eq 1 ]]; then
-	    echo "[DEBUG] Check children of $checkurl"
+		log "[DEBUG] Check children of $checkurl"
 	fi
 
 	curl -s ${checkurl} > ${tmpdir}/composite.xml
 	if [[ ! -f ${tmpdir}/composite.xml ]] || [[ $(egrep "404 Not Found" ${tmpdir}/composite.xml) ]]; then
-		echo "[ERROR] Could not read ${checkurl} ! "
-		rm -fr ${tmpdir}
-		exit 1
-	fi
-	if [[ $debug -eq 1 ]]; then
-		echo "[DEBUG] ${tmpdir}/composite.xml"
-		echo "[DEBUG] ------------"
-		cat ${tmpdir}/composite.xml
-		echo "[DEBUG] ------------"
-	fi
-	cat ${tmpdir}/composite.xml | egrep "<child " | egrep -v "<\!--" | egrep "location" \
-	  | sed -e "s#.*<child location=[\'\"]\+\([^\'\"]\+\)[\'\"]\+.*\/>.*#\1#" > ${tmpdir}/composite.txt
-	urls=$(cat ${tmpdir}/composite.txt)
-	if [[ $debug -eq 1 ]]; then
-		echo "[DEBUG] Got urls (in ${tmpdir}/composite.txt ):"
-		echo "[DEBUG] ------------"
-		cnt=0
-		for url in $urls; do 
-			(( cnt = cnt + 1 ))
-			echo "[$cnt] $url"
-		done
-		echo "[DEBUG] ------------"
-	fi
-	rm -f ${tmpdir}/composite.txt ${tmpdir}/composite.xml
-
-	countItems "${urls}"
-	(( numUrls = numUrls + itemCount ))
-
-	newurls=""
-    for a in ${urls}; do
-        if [[ ${a} != "http"* ]]; then # relative path
-            a=${checkurl%/composite*xml}/${a}
-        fi
-	    newurls="${newurls} ${a}"
-    done
-    urls=${newurls}
-
-	if [[ $debug -eq 1 ]]; then
-		echo "[DEBUG] Got urls:"
-		echo "[DEBUG] ------------"
-		cnt=0
-		for url in $urls; do 
-			(( cnt = cnt + 1 ))
-			echo "[$cnt] $url"
-		done
-		echo "[DEBUG] ------------"
-	fi
- 
-    for a in ${urls}; do
-		((num = num + 1 ))
-
-		indent=""
-		if [[ $indentNum -gt 0 ]]; then
-			for i in `seq 1 ${indentNum}`; do
-				indent="${indent} >"
-			done
+		# check if the folder exists (but isn't a composite)
+		curl -s ${checkurl%/composite*.xml} > ${tmpdir}/index.html
+		if [[ ! -f ${tmpdir}/index.html ]] || [[ $(egrep "404 Not Found" ${tmpdir}/index.html) ]]; then
+			logerr "[ERROR] Could not read ${checkurl} ! "
+			rm -fr ${tmpdir}
+			exit 1
+		else
+			checkurl=${checkurl%/composite*.xml}
+			log "[WARNING] ${checkurl} is not a composite site ! "
 		fi
-        logn "{${checkurlnum}/${checkurltot}} [${num}/${numUrls}]${indent} ${a} : "; stat=$(curl -I -s ${a} | egrep "404 Not Found")
-        if [[ ! $stat ]]; then 
-        	log "${green}OK${norm}"
-        	let OK+=1
-			stat=$(curl -I -s ${a}/compositeContent.xml | egrep "404 Not Found")
-			if [[ ! $stat ]]; then # exists
-				(( indentNum = indentNum + 1 ))
-				checkCompositeXML ${a}/compositeContent.xml ${indentNum}
-				(( indentNum = indentNum - 1 ))
-    		fi
-	    else 
-        	if [[ $quiet == 0 ]]; then 
-	        	logerr "{${checkurlnum}/${checkurltot}} [${num}/${numUrls}]${indent} ${a} : " "${red}NO${norm} \n$(curl -I -s ${a})"
-	        else
-    	    	logerr "{${checkurlnum}/${checkurltot}} [${num}/${numUrls}]${indent} ${a} : " "${red}NO${norm}"
-    	    fi
-        	let notOK+=1
-        fi
-    done
+	fi
+	if [[ -f ${tmpdir}/composite.xml ]]; then
+		if [[ $debug -eq 1 ]]; then
+			log "[DEBUG] ${tmpdir}/composite.xml"
+			log "[DEBUG] ------------"
+			cat ${tmpdir}/composite.xml
+			log "[DEBUG] ------------"
+		fi
+		cat ${tmpdir}/composite.xml | egrep "<child " | egrep -v "<\!--" | egrep "location" \
+		  | sed -e "s#.*<child location=[\'\"]\+\([^\'\"]\+\)[\'\"]\+.*\/>.*#\1#" > ${tmpdir}/composite.txt
+		urls=$(cat ${tmpdir}/composite.txt)
+		if [[ $debug -eq 1 ]]; then
+			log "[DEBUG] Got urls (in ${tmpdir}/composite.txt ):"
+			log "[DEBUG] ------------"
+			cnt=0
+			for url in $urls; do 
+				(( cnt = cnt + 1 ))
+				log "[$cnt] $url"
+			done
+			log "[DEBUG] ------------"
+		fi
+		rm -f ${tmpdir}/composite.txt ${tmpdir}/composite.xml
+	elif [[ -f ${tmpdir}/index.html ]]; then
+		urls="${checkurl}"
+		# TODO parse the index page for links to validate?
+	else
+		urls="${checkurl}"
+	fi
+
+	if [[ ${urls} ]]; then
+		countItems "${urls}"
+		(( numUrls = numUrls + itemCount ))
+
+		newurls=""
+		for a in ${urls}; do
+			if [[ ${a} != "http"* ]]; then # relative path
+				a=${checkurl%/composite*xml}/${a}
+			fi
+			newurls="${newurls} ${a}"
+		done
+		urls=${newurls}
+
+		if [[ $debug -eq 1 ]]; then
+			log "[DEBUG] Got urls:"
+			log "[DEBUG] ------------"
+			cnt=0
+			for url in $urls; do 
+				(( cnt = cnt + 1 ))
+				log "[$cnt] $url"
+			done
+			log "[DEBUG] ------------"
+		fi
+	 
+		for a in ${urls}; do
+			((num = num + 1 ))
+
+			indent=""
+			if [[ $indentNum -gt 0 ]]; then
+				for i in `seq 1 ${indentNum}`; do
+					indent="${indent} >"
+				done
+			fi
+			logn "{${checkurlnum}/${checkurltot}} [${num}/${numUrls}]${indent} ${a} : "; stat=$(curl -I -s ${a} | egrep "404 Not Found")
+			if [[ ! $stat ]]; then 
+				log "${green}OK${norm}"
+				let OK+=1
+				stat=$(curl -I -s ${a}/compositeContent.xml | egrep "404 Not Found")
+				if [[ ! $stat ]]; then # exists
+					(( indentNum = indentNum + 1 ))
+					checkCompositeXML ${a}/compositeContent.xml ${indentNum}
+					(( indentNum = indentNum - 1 ))
+				fi
+			else 
+				if [[ $quiet == 0 ]]; then 
+					logerr "{${checkurlnum}/${checkurltot}} [${num}/${numUrls}]${indent} ${a} : " "${red}NO${norm} \n$(curl -I -s ${a})"
+				else
+					logerr "{${checkurlnum}/${checkurltot}} [${num}/${numUrls}]${indent} ${a} : " "${red}NO${norm}"
+				fi
+				let notOK+=1
+			fi
+		done
+	fi
 }
 
 checkurlnum=0
@@ -187,7 +204,7 @@ for checkurl in ${checkurls}; do
 
 	tmpdir=${WORKSPACE}/${0##*/}_tmp; mkdir -p $tmpdir
 	if [[ $debug -eq 1 ]]; then
-		echo "[DEBUG] tmpdir: ${tmpdir}"
+		log "[DEBUG] tmpdir: ${tmpdir}"
 	fi
 
 	stat=$(curl -I -s ${checkurl} | egrep "404 Not Found")
@@ -197,9 +214,15 @@ for checkurl in ${checkurls}; do
 		numUrls=0
 		checkCompositeXML ${checkurl} 0
 	else
-		logerr "" "[ERROR] Could not read ${checkurl} ! "
-		rm -fr ${tmpdir}
-		exit 1
+		curl -s ${checkurl%composite*.xml} > ${tmpdir}/index.html
+		if [[ ! -f ${tmpdir}/index.html ]] || [[ $(egrep "404 Not Found" ${tmpdir}/index.html) ]]; then
+			logerr "[ERROR] Could not read ${checkurl} ! "
+			rm -fr ${tmpdir}
+			exit 1
+		else
+			checkurl=${checkurl%/composite*.xml}
+			log "[WARNING] ${checkurl} is not a composite site: ${blue}OK${norm}"
+		fi
 	fi
 	rm -fr ${tmpdir}
 done
